[Profiler] Fix UBsan Job (#6380)
## Summary of changes

Fix the UBSAN job.

## Reason for change

A recent unit test was added. This test builds events by writing the
field in raw memory (like the CLR does). But the alignment is not taken
care of (like the CLR) and UBSAN shouts about.

## Implementation details

Since it's a test, we just disable UBSAN on the function (in the test)
that creates the event.