Remove old generated public APIs (#6376)
## Summary of changes

Delete the Public API Generator

## Reason for change

We added this generator to reduce the amount of boilerplate you need to
write to correctly record telemetry for public APIs. However, with the
move to Datadog.Trace.Manual, these APIs are completely unused, so we
may as well remove the complexity and duplication.

## Implementation details

- Delete the `[GeneratePublicApi]` generator
- Keeping the `[PublicApi]` attribute for now - that will be cleared up
in a separate PR
- Rename the previously-decorated properties that _were_ called
`SomePropInternal` to `SomeProp`
- We still have some "leftover" `Internal` suffixes, will tidy those up
separately
- Main changes are in `TracerSettings`, `ExporterSettings`,
`IntegrationSettings` + their immutable counterparts

## Test coverage

Covered by existing tests

## Other details

I noticed that there were some properties that we were getting public
API telemetry for which we kind of lost in the move to
Datadog.Trace.Manual: `SpanContext.Parent`, `SpanContext.ParentId`,
`SpanContext.ServiceName`. I don't _think_ that actually matters, as
these can't actually be accessed any more for technical reasons...

Part of stack 
- https://github.com/DataDog/dd-trace-dotnet/pull/6370
- https://github.com/DataDog/dd-trace-dotnet/pull/6376 👈 This PR
- https://github.com/DataDog/dd-trace-dotnet/pull/6385
- https://github.com/DataDog/dd-trace-dotnet/pull/6386
- https://github.com/DataDog/dd-trace-dotnet/pull/6397
- https://github.com/DataDog/dd-trace-dotnet/pull/6399
- https://github.com/DataDog/dd-trace-dotnet/pull/6405