[CI Visibility] Set `TracerSettings` using a configuration source instead of mutating (#6399)
## Summary of changes

Change how CI Visibility creates its `TracerSettings` object to avoid
mutating settings afterwards

## Reason for change

This stack of changes is about removing duplication (among other things)
in the `TracerSettings` etc related objects. This is _partially_
required in order to remove the "snapshot generator" complexity that was
removed in https://github.com/DataDog/dd-trace-dotnet/pull/6370. Given
`TracerSettings` are not exposed publicly in Datadog.Trace, we want to
move to doing a "one shot" configuring of them, ultimately so that we
can make the object immutable (and remove `ImmutableTracerSettings` and
friends).

## Implementation details

CI Visibility is one of the few places where we mutate settings _after_
creating the `TracerSettings`. This is mostly because there's additional
logic that CI Visibility wants to perform.

Ultimately, the "solution" to that issue in this PR is to move that
logic into the `TracerSettings` constructor. I'm not entirely happy
about it, but it's the only approach I could find that works.

- Add a "dummy" configuration value to a configuration source when
creating the `TracerSettings`. This is used purely as a "switch" to say
"we're in CI Visibility mode".
- We absolutely could just pass this in as a constructor parameter. I
went that route first and then backed away, but can totally be swayed.
- Add an additional configuration source to update settings that we want
to change in CI Vis (e.g. logs injection).
- In the `TracerSettings` ctor, add an additional ignored URL when in CI
Vis mode, modify the default service name (if required) and add an
additional GlobalTag.

## Test coverage

I'd love to have some, but the CI Visibility configuration is kinda all
over the place, so if you have any pointers @tonyredondo I'm all ears...

## Other details

Part of stack 
- https://github.com/DataDog/dd-trace-dotnet/pull/6370
- https://github.com/DataDog/dd-trace-dotnet/pull/6376
- https://github.com/DataDog/dd-trace-dotnet/pull/6385
- https://github.com/DataDog/dd-trace-dotnet/pull/6386
- https://github.com/DataDog/dd-trace-dotnet/pull/6397 
- https://github.com/DataDog/dd-trace-dotnet/pull/6399 👈 This PR
- https://github.com/DataDog/dd-trace-dotnet/pull/6400  
- https://github.com/DataDog/dd-trace-dotnet/pull/6405
- https://github.com/DataDog/dd-trace-dotnet/pull/6408