add the filepath we check for datadog.json to the output of diagnostic tool (#6273)
outputs `configurationFileName` in TryLoadJsonConfigurationFile and we
print it

## Summary of changes
add the directory where we look for `datadog.json` in the tracer logs
and in the diagnostic tool
## Reason for change
https://datadoghq.atlassian.net/browse/APMLP-297
Customer confusion over why their configuration isn't picked up. 

## Implementation details
Log where the directory we check the json for is.

## Test coverage
Need to create a test to check this programmatically in the future, but here's the
results running it manually on IIS/Process (IIS check calls the process
check)
datadog.json read: 

![image](https://github.com/user-attachments/assets/8614c834-e08d-47af-8135-f5a10000ca4d)
datadog.json not read: 

![image](https://github.com/user-attachments/assets/c49fe373-47da-47e6-b604-29d673d93533)


## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->