[ASM] RASP: Command injection vulnerability implementation (#6323)
## Summary of changes

This PR contains the implementation of the RASP command injection
vulnerability. Previously, the shell injection vulnerability was
implemented, which is similar but only occurs the flag UseShellExecute
is set to false in method process.start().

The definition of this new vulnerability can be found here:
https://docs.google.com/document/d/1DDWy3frMXDTAbk-BfnZ1FdRwuPx6Pl7AWyR4zjqRFZw/edit?tab=t.0#heading=h.giijrtyn1fdx

## Reason for change

It's required for RASP.

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->

---------

Co-authored-by: Anna <anna.yafi@datadoghq.com>
Co-authored-by: Andrew Lock <andrew.lock@datadoghq.com>