[ASM] Fix IAST weak random vulnerability error (#6432)
## Summary of changes

When running an app out of a request context, we were not closing the
scope after creating a vulnerability span in the case of an IAST weak
random vulnerability. That would make the tracer to stop sending spans
and would store them forever, taking an increasing amount of memory.

Basically, in the method close_span of the class traceContext, the
variable _openSpans would never reach 0, so the array _spans would never
be empty and would keep storing spans forever.

The problem is that, in the System.Random aspect, we were not closing
the scope/span, which we were actually doing in other aspects such as
the weak hash aspect.

A little refactor has been done. Now, by default, we close an IAST
vulnerability span right after creation in non request scenarios. In the
call target instrumentations, we already handle the closing of the span,
so the IAST instrumentations that can be launched in non request
scenarios will keep the span closing that was already implemented.

A filter has been added to avoid instrumenting the Grpc.Net.Client
assembly. We were reporting a vulnerability inside this library that was
adding noise.

In the deduplication tests, a new random.Next call has been added.
Before the implemented changes, this test would receive no spans from
the tracer after adding the Random.Next call.

## Reason for change

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->