[dd-dotnet] Catch error when failing to read env vars on IIS (#6430)
## Summary of changes

Catch the error and log a warning when the diagnostic tool fails to read
custom environment variables on an IIS application pool.

## Reason for change

Versions of IIS prior to 10 didn't have an `<environmentVariables>`
node.

## Implementation details

Just catching the exceptions. And disabling rule SA1108 because stylecop
doesn't like my comment.

## Test coverage

I tried replacing `environmentVariables` with an attribute that doesn't
exist and it works.