Data Streams Monitoring support in Kinesis (#6428)
## Summary of changes

* Added Data Streams Monitoring checkpoints to existing
autoinstrumentation of Kinesis PutRecord[s][Async]
* Added autoinstrumentation to Kinesis GetRecords[Async]

## Reason for change

Data Streams Monitoring was not supported for Kinesis, this adds that
functionality.

## Implementation details

The existing producer instrumentation only works when StreamName is
specified, not when StreamARN is specified. These changes leave that
behavior unchanged at present.

The new consumer instrumentation is able to infer a StreamName from the
[mandatory] StreamARN so that the producer and consumer can use the same
naming system. The reverse would not be possible, and changing producer
instrumentation to use ARNs instead would be a potentially-breaking
change to existing users of the functionality so is outside the scope of
this feature addition.