Add support for `[DuckPropertyOrField]` (#6463)
## Summary of changes

Allows you to specify that a duck type member can match a property _or_
a field - whatever's available.

## Reason for change

In the recent version of RabbitMQ, they changed some properties into
identically-named fields. Without this change, we'd have to duplicate
the whole duck chain hierarchy.

## Implementation details

If the `DuckKind` is `PropertyOrField`, and we fail to find a property,
then try to look for a matching field instead.

## Test coverage

Added some simple tests, that it works as expected, but I don't think we
need to go as far as duplicating the extensive suite we have for
Property or Field matching.

## Other details

Required for RabbitMQ v7 support