K8s Lib Injection tests: run on a matrix (#6458)
## Summary of changes
This PR run the K8s Lib injection tests on a matrix.

## Reason for change
This change is necessary because we are going to change the "one
pipeline" and these changes could break dd-trace-js pipeline. We need to
merge this PR before merging the "one pipeline".
it is the first PR to do this adaptation step by step.

## Other details
Coming PRs to apply the change:

system-tests PR changing the K8s lib injection tests:
https://github.com/DataDog/system-tests/pull/3689
one pipeline PR:
https://github.com/DataDog/libdatadog-build/pull/59/files
Final PR on dd-trace-dotnet:
https://github.com/DataDog/dd-trace-dotnet/pull/6451/files