Fix the build and some versions (#6466)
## Summary of changes

Fix the build and the version-bump action

## Reason for change

In https://github.com/DataDog/dd-trace-dotnet/pull/6457, @vandonr
spotted that we accidentally, incorrectly, "bumped" the required version
of cmake.


![image](https://github.com/user-attachments/assets/5843c1db-2dad-439e-8ed8-eeeb2d05def5)

Additionally, the new calltargets_generated.g.cpp file _also_ now embeds
the version, so we need to either update that, or stop embedding the
version.

@daniel-romano-DD is looking into the latter approach
[here](https://github.com/DataDog/dd-trace-dotnet/pull/6464), so this PR
is just a fix to unblock PRs.

## Implementation details

- Revert the broken CMakeLists.txt
- Fix the broken generated file
- Be more discerning in our CMakeLists updates

## Test coverage

Tested the version bump routine locally, and confirmed it works as
expected now.

## Other details

It would be _really_ nice if the version bumping just required bumping
one or two values... That _might_ be somewhat feasible in the future,
but for now...