[Crashtracking] Add a setting to disable crashtracking filtering (#6452)
## Summary of changes

Add a `DD_CRASHTRACKING_FILTERING_ENABLED` environment variable. When
set to 0, crashtracking will send all crash reports to telemetry,
regardless of whether they're marked as suspicious or not.

## Reason for change

It is not always easy for customers to send crash dumps. This can be
used as an alternative solution in some cases. We _may_ also be able to
use it to get information on some crashes in the CI without having to
download the coredump.

## Test coverage

Added a test to confirm a report is received in a non-datadog crash when
setting `DD_CRASHTRACKING_FILTERING_ENABLED=0`