[IAST] Fix recursive custom attribute crash (#6470)
## Summary of changes
Break recursion when discovering custom attributes of a method and that
method was the constructor of an attribute decorated with the same
attribute.

## Reason for change
A stack overflow crash would happen if a constructor of an attribute was
decorated with the same attribute.

## Implementation details
We only want type name of the method in order to retrieve the attributes
applied. Instead of retrieving the MemberRefInfo class, which read the
custom attributes on the constructor, we now query the profiling API for
the typedef of the method, and get its name directly.

## Test coverage
A unit test has been provided. This test crashed before the fix.

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->