Parallelize unit tests (#6483)
## Summary of changes

Parallelize the unit test runs in CI

## Reason for change

As we test more frameworks, the runs take longer. We can easily
parallize by TFM (e.g. we allow running a specific TFM locally), so this
overall speeds up the unit test execution. As we (attempt) to improve
the speed of other stages, this allows gaining confidence earlier.
Additionally, in the case of flake, re-running a stage takes less time
when each TFM is parallelized.

## Implementation details

Generate the matricies, use them in CI

## Test coverage

Did a test run, looks OK to me

## Other details

Reduces the wall time from this:


![image](https://github.com/user-attachments/assets/e62b4743-5726-4309-a441-c3f5af144c2e)

![image](https://github.com/user-attachments/assets/75733f99-8aa9-49ca-9ff5-69cb6bf98644)

to this:


![image](https://github.com/user-attachments/assets/2a4fcb56-0c34-416b-8070-9fc9f5bfa9c9)