Support relative path in `DD_DOTNET_TRACER_HOME` (#6434)
## Summary of changes

Resolve the path read from `DD_DOTNET_TRACER_HOME` to an absolute path
before trying to load an assembly from there.

## Reason for change

When writing a sample app and adding `Datadog.Trace.Bundle` as a
dependency, it can be natural to set those variables to something like
this :

`CORECLR_PROFILER_PATH=./datadog/osx/Datadog.Trace.ClrProfiler.Native.dylib`
`DD_DOTNET_TRACER_HOME=./datadog`

In recent versions of .NET, `CORECLR_PROFILER_PATH ` supports a relative
path, but `DD_DOTNET_TRACER_HOME` doesn't and never has, and throws an
`ArgumentException` [as
documented](https://learn.microsoft.com/en-us/dotnet/api/system.runtime.loader.assemblyloadcontext.loadfromassemblypath?view=net-9.0)
when you try this with v3 of the packages.

However, in v2 of the library, if you referenced `Datadog.Trace.Bundle`
or `Datadog.Trace`, the value of `DD_DOTNET_TRACER_HOME` was _never
actually used_. That meant that it _appeared_ as though relative paths
worked.

In actual fact, what was happening is that referencing one of these
NuGet packages caused the `Datadog.Trace.dll` file to be copied to the
`bin`/`publish` output folder. As this dll was always next to the
`.exe`, it was loaded automatically by the runtime, and never invoked
the `AssemblyResolve` event code in the managed loader to load from
monitoring home.

In v3, we no longer ship `Datadog.Trace.dll` in the NuGet packages,
instead shipping `Datadog.Trace.Manual.dll`. That means that the
`DD_DOTNET_TRACER_HOME` variable _is_ used with v3, even when you
reference the bundle, and gives the _appearance_ that "relative paths
are broken". To work around that, this PR adds support for relative
paths.

To be clear, we **still don't recommend using relative paths for these
variables** in general. The difficulty in understanding what the path is
_relative to_, means that it will be often be more hassle than it's
worth. That said, it can make sense in the bundle scenario described
above, hence this PR.

## Implementation details

Call `GetFullPath()` on the value provided in `DD_DOTNET_TRACER_HOME`.
This uses a native call (e.g. `Interop.Sys.GetCwd() on linux,
GetFullPathNameW()` on Windows) to determine the full path given a
(potentially) relative path.

## Test coverage

- Added a couple of tests to confirm relative paths work.
- Added a shim implementation of `Path.GetRelativePath()` for .NET
Framework based on the .NET Core implementation.

## Other details

Interestingly, getting the _correct_ relative path in the integration
tests is a massive pain. The problem is that the relative path you need
is _not_ the path relative to the executable, it's relative to "the
current directory", which could be anywhere, depending on how you run
the application. In the integration tests, for example, the current
directory is the directory containing the _integration test executable_
💀

---------

Co-authored-by: Andrew Lock <andrew.lock@datadoghq.com>