Disable continuous profiler in GenerateDumpIfDbgRequested (#6497)
## Summary of changes

Disable the continuous profiler in `GenerateDumpIfDbgRequested`

## Reason for change

`GenerateDumpIfDbgRequested` has been flaky for a while now and I've
been unable to find the root cause. I suspect the failure might be
caused by some weird interaction with the continuous profiler, and the
CP isn't actually needed for this test, so I'm disabling it to see if
the failure disappears.