Expand number of files that require thorough testing (#6504)
## Summary of changes

Try to catch more integration issues in PRs

## Reason for change

We don't run all TFMs on all PRs, because it increases the load on CI,
and doesn't add a huge amount of value. Unfortunately, that means we
also occasionally merge PRs that break an untested-in-PR TFM which then
fails on master. We attempt to mitigate that, but running all TFMs for
"high risk of breakage" PRs. This increases the scope of what
constitutes "high risk"

## Implementation details

Previously, we were only counting PRs which change the calltarget or
callsite integration points. This PR updates that to include the
_implementations_ too

## Test coverage

Tested locally to confirm it works as expected