Log an error message when the profiler is loaded multiple times (#6503)
## Summary of changes

Log an error message when the profiler is loaded multiple times.

## Reason for change

Despite mentioning it in our documentation, we still regularly have
escalations caused by the usage of .NET Framework and .NET Core in the
same pool. Adding an error log will make diagnostic easier.

We could also entirely bail out when we detect multiple initialization,
but I'm afraid we might break existing customers.

## Implementation details

Using the `profiler` field to find out if the profiler was initialized.
It means we won't detect double-initialization if the first
initialization failed midway through, but I don't think we really care.

## Test coverage

Tested on my machine.