[ASM] Fix access violation exception when reading WAF addresses (#6510)
## Summary of changes

There are some errors detected in real customers that seem to be related
to the method Waf.GetKnownAddresses() All the following errors seem to
be caused by the same issue: memory corruption.

The errors detected include:

```
exception:Type: System.AccessViolationException
Message: Attempted to read or write protected memory. This is often an indication that other memory is corrupt.

Unhandled Exception: System.ObjectDisposedException: Cannot access a disposed object.
Object name: 'UnmanagedMemoryPool'.

System.NullReferenceException
   at System.SpanHelpers.IndexOfNullByte(Byte* searchSpace)
   at System.String.Ctor(SByte* value)
   at System.Runtime.InteropServices.Marshal.PtrToStringAnsi(IntPtr ptr)
   at Datadog.Trace.AppSec.Waf.NativeBindings.WafLibraryInvoker.GetKnownAddresses(IntPtr wafHandle)
```
   
The stack traces suggest that the problem would be originated in the WAF
dll:

```
System.AccessViolationException
        C:\\Windows\\SYSTEM32\\ntdll.dll!<unknown>+a38ee
        C:\\Program Files\\Datadog\\.NET Tracer\\win-x64\\ddwaf.dll!<unknown>+4f34e
        Datadog.Trace.dll!.IL_STUB_PInvoke
        Datadog.Trace.dll!Datadog.Trace.AppSec.Waf.NativeBindings.WafLibraryInvoker.GetKnownAddresses
        Datadog.Trace.dll!Datadog.Trace.AppSec.Security.UpdateActiveAddresses

```

A new unit test was created. This unit test launches 100 parallel calls
to the WAF GetKnownAddresses method. The test was able to reproduce the
error, crashing sometimes the testhost.exe process.

It seems that the WAF builds a cache when receiving the first
GetKnownAddresses request. Parallel requests seem to be causing a
concurrency issue. A lock would prevent this issue. It would seem that
we would only need to protect the GetKnownAddresses calls during the
first call, but, when doing that, the unit test still failed. When
locking all the calls to the GetKnownAddresses method, tests passed
consistently. That's why a WriteLock has been used.

Anyway, this method is only called when there are RC changes, so the
performance impact of the lock should be minimal.

## Reason for change

We were getting some of these exceptions logs from real customers.

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->