Add support for more combinations of tags on telemetry metrics (#6429)
## Summary of changes

Allow passing in more tag values to telemetry metrics

## Reason for change

@tonyredondo requested it as it would make some of the combinatorial tag
requirements for ci visibility easier.

## Implementation details

Generalised the current behaviour (that limited it to 2 metrics)

> [!WARNING]
> The original 2 metrics limitation was _partly_ there to avoid the
combinatorial explosion you can get with more metrics. Be aware that we
_always_ generate _all_ combinations of tags, so if only _certain_
combinations of tags should be set together, then you should use the
existing capabilities of grouping multiple tags.

During the factoring, I was careful to keep the generated code
consistent. We can sweep through and simplify the generator (by allowing
trivial renaming of variables etc) later, but didn't want to muddy the
waters for review in this PR.

## Test coverage

Added unit tests for the source generator. We should carefully test and
review the first _real_ usage of the multiple-generics, to confirm
there's no unexpected issues.