Include extra tags in error logs (#6364)
## Summary of changes

Includes details about products enabled in the telemetry error logs

## Reason for change

Should make it easier to categorize errors knowing e.g. whether the
error came from a ci-vis app etc.

## Implementation details

The tags are written to _every_ log message, so we need to keep the tags
small as possible. I limited it to just 5 tags for now:

- `ci` - is CI Visibility enabled
- `asm` - is ASM enabled
- `prof` - is profiling enabled
- `dyn` - is dynamic instrumentation enabled
- `gcp`/`aws`/`aas`/azf` - are we running in GCP/lambda/AAS/Azure
functions (with mini agent)

I _think_ those cover the broadest categories for now. I opted for
adding the enabled tag with a `1`/`0` so we could explicitly see whether
it was set or not (apart from the cloud functions which are only added
when they apply).

Alternatively, we could _only_ add the tag if it's enabled (like I do
for the "cloud" tag). That saves bytes, at the expense of some
ambiguity: if the tags aren't there does that mean they definitely were
none enabled, or the tagging didn't happen? Happy to switch though if
people have a preference.

## Test coverage

Unit tests. Will do a manual test to make sure there's no issues with
the intake accepting them.

## Other details

We currently weren't explicitly recording whether CI Vis was enabled, so
I added a method to call to "record" the settings. This isn't _really_
required given we can only get the enabled flag via a static property,
but it _potentially_ sets us up a bit more cleanly for refactoring
later, and is closer to how other products work. I'm happy to be
persuaded to remove that commit though

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->