Enable dependabot for Azure Functions sample (#6527)
## Summary of changes

Try to enable dependabot for the Azure Functions v2 sample

## Reason for change

In https://github.com/DataDog/dd-trace-dotnet/pull/6524 we had to remove
our "normal" version bumping work, because the Azure Functions samples
fundamentally don't work well with the `dotnet msbuild` approach. This
PR attempts to give us some of that visibility back by explicitly
configuring the azure functions sample in dependabot

## Implementation details

Added the sample directly to the yml. I think this is the most important
one to track, as I think it's the only currently developed scenario that
we expect to see updates.

## Test coverage

Can't easily test the dependabot changes unfortunately AFAIK